Move sensor rotation screen into screen package
