+++ app/src/main/java/com/skul/yuriy/composeplayground/feature/sensorRotation/SensorRotationActivity.kt
@@ -7,13 +7,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.SystemBarStyle
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -24,29 +18,15 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableFloatStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.TransformOrigin
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
-import com.skul.yuriy.composeplayground.R
-import com.skul.yuriy.composeplayground.feature.sensorRotation.sensor.AccelerometerRotationAngleSource
-import com.skul.yuriy.composeplayground.feature.sensorRotation.sensor.OrientationEventRotationAngleSource
-import com.skul.yuriy.composeplayground.feature.sensorRotation.sensor.RotationAngleSourceType
+import com.skul.yuriy.composeplayground.feature.sensorRotation.screen.SensorRotationScreen
 import com.skul.yuriy.composeplayground.ui.theme.ComposePlaygroundTheme
 
 class SensorRotationActivity : ComponentActivity() {
@@ -93,107 +73,6 @@ class SensorRotationActivity : ComponentActivity() {
         super.onStop()
     }
 }
-
-@Composable
-private fun SensorRotationScreen(
-    onNavUp: () -> Unit
-) {
-    val tiltAngle = rememberRotationAngle()
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.White)
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(4.dp)
-                .border(width = 1.dp, color = Color.Red)
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .graphicsLayer {
-                        rotationZ = tiltAngle
-                        transformOrigin = TransformOrigin.Center
-                    }
-            ) {
-                Canvas(
-                    modifier = Modifier.fillMaxSize()
-                ) {
-                    val centerY = size.height / 2f
-                    drawLine(
-                        color = Color.Black,
-                        start = Offset(x = 0f, y = centerY),
-                        end = Offset(x = size.width, y = centerY),
-                        strokeWidth = 2.dp.toPx()
-                    )
-                }
-
-                Text(
-                    modifier = Modifier.align(Alignment.Center),
-                    text = "Tilt: ${"%.1f".format(tiltAngle)}°",
-                    color = Color.Black,
-                    fontWeight = FontWeight.Bold
-                )
-            }
-
-            // dont delete - прост опока он не нужен
-
-//            SensorRotationTopAppBar(
-//                modifier = Modifier
-//                    .align(Alignment.TopCenter)
-//                    .padding(top = rememberStatusBarHeight()),
-//                title = stringResource(R.string.sensor_rotation_demo),
-//                onNavUp = onNavUp
-//            )
-//
-//            Text(
-//                modifier = Modifier.align(Alignment.Center),
-//                text = stringResource(R.string.sensor_rotation_demo_placeholder),
-//                color = Color.Black
-//            )
-        }
-    }
-}
-
-@Composable
-private fun rememberRotationAngle(
-    sourceType: RotationAngleSourceType = DEFAULT_ROTATION_ANGLE_SOURCE             // old
-//    sourceType: RotationAngleSourceType = ALTERNATIVE_ROTATION_ANGLE_SOURCE           //new
-): Float {
-    val context = LocalContext.current
-    val angleSource = remember(context, sourceType) {
-        when (sourceType) {
-            RotationAngleSourceType.Accelerometer -> AccelerometerRotationAngleSource(context)
-            RotationAngleSourceType.OrientationEventListener -> {
-                OrientationEventRotationAngleSource(context)
-            }
-        }
-    }
-    var tiltAngle by remember { mutableFloatStateOf(0f) }
-    val currentOnAngleChanged by rememberUpdatedState<(Float) -> Unit> { tiltAngle = it }
-
-    DisposableEffect(angleSource) {
-        angleSource.start { angle ->
-            currentOnAngleChanged(angle)
-        }
-        onDispose(angleSource::stop)
-    }
-
-    return tiltAngle
-}
-private val DEFAULT_ROTATION_ANGLE_SOURCE = RotationAngleSourceType.Accelerometer
-private val ALTERNATIVE_ROTATION_ANGLE_SOURCE = RotationAngleSourceType.OrientationEventListener
-
-
-
-
-
-
-
-
-
